segundo cambio con valores enteros

+++ sumanumeros.py
@@ -1,4 +1,6 @@
-num1=float(input("ingrese un numero: "))
-num2=float(input("ingrese un numero: "))
+num1=int(input("ingrese un numero: "))
+num2=int(input("ingrese un numero: "))
+
 sumar=num1+num2 
+
 print("la suma de dos numeros es: ",sumar)
